See if React Flow works with server components.
Answer: it does not. ;(

I added two examples, `app/flow-csr` and `app/flow-ssr`.

The CSR example contains a single server component, that renders a client component that
uses React Flow. This works. As expected, it first renders an empty initial page (we could
render a skeleton loader instead), and then browser evalutes the JS and renders the
flow. If you disable Javascript you see a blank screen.

The SSR example blows up because React Flow tries to use a React Context in a server
component, which is not allowed:

> createContext only works in Client Components. Add the "use client" directive at the top
of the file to use it. Read more:
https://nextjs.org/docs/messages/context-in-server-component

![](https://cdn.zappy.app/8602ef9f00626bfbd8cf178c5c35d426.png)