See if react flow works with old school SSR (no server components).
Answer: it does not.

I added two new pages, pages/simple-ssr-page and pages/flow-page.

The first just renders some simple text SSR. It works. Turn off JS, and it still works.

The second tries renders a flow SSR in a page. With JS turned off it renders a blank
page. With JS turned on, it renders the flow just fine, but CSR on the client, after the
JS has loaded.